refactor: extract drawdown benchmark comparison
